Added tests, reorganized folder structure for tests, updated NavLink to include default theme value,
