+++ src/tests/Header.test.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import Header from '../components/Header';
+
+describe('Header', () => {
+  it('renders correctly', () => {
+    const component = renderer.create(
+      <Header />,
+    );
+
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
